rename redux folder to appRedux, update imports
Webpack looks for import paths in the node_modules/redux folder instead of the project structure.
Renamed to avoid redux name conflict and having to use ../../../redux
