+++ frontend/src/appRedux/books/operations.tsx
