+++ frontend/src/appRedux/books/selectors.tsx
